+++ lib_xcore_math/src/arch/ref/dct/dct.c
@@ -18,7 +18,7 @@ void vect_s32_flip(
     const int32_t b[],
     const unsigned length)
 {
-  for(int k = 0; k < (length>>1); k++){
+  for(size_t k = 0; k < (length>>1); k++){
     int32_t x = b[k];
     int32_t y = b[length-1-k];
     a[k] = y;
@@ -35,7 +35,7 @@ headroom_t dct_adsb_s32(
     const int32_t dct_lut[])
 {
   const unsigned length = 8*chunks;
-  for(int k = 0; k < length; k++){
+  for(size_t k = 0; k < length; k++){
     int64_t h = head[k];
     int64_t t = tail[k];
     int64_t s = h + t;
@@ -56,7 +56,7 @@ void dct_deconvolve_s32(
     const unsigned res_length)
 {
   int32_t prev = D[0] >> 1;
-  for(int k = 0; k < (res_length >> 1); k++){
+  for(size_t k = 0; k < (res_length >> 1); k++){
     int32_t b = B[k];
     int32_t d = D[k];
     res[2*k] = b;
@@ -71,7 +71,7 @@ void dct6_forward(
 {
   const unsigned DCT_N = 6;
   int32_t tmp[6];
-  for(int k = 0; k < DCT_N; k++)
+  for(size_t k = 0; k < DCT_N; k++)
     tmp[k] = vlmaccr32(0, x, &dct6_matrix[DCT_N-1-k][0]);
   memcpy(y, tmp, sizeof(tmp));
 }
@@ -83,7 +83,7 @@ void dct8_forward(
 {
   const unsigned DCT_N = 8;
   int32_t tmp[8];
-  for(int k = 0; k < DCT_N; k++)
+  for(size_t k = 0; k < DCT_N; k++)
     tmp[k] = vlsat32(vlmaccr32(0, x, &dct8_matrix[DCT_N-1-k][0]), 3);
   memcpy(y, tmp, sizeof(tmp));
 }
