+++ lib_xcore_math/src/arch/ref/vect_macc.c
@@ -22,7 +22,7 @@ headroom_t vect_s16_macc(
     const right_shift_t acc_shr,
     const right_shift_t bc_shr)
 {
-    for(int k = 0; k < length; k++){
+    for(size_t k = 0; k < length; k++){
         acc[k] = vlashr16(acc[k], acc_shr);
         const vpu_int16_acc_t tmp = vlmacc16(0, b[k], c[k]);
         acc[k] = vladd16(acc[k], vlsat16(tmp, bc_shr));
@@ -40,7 +40,7 @@ headroom_t vect_s16_nmacc(
     const right_shift_t bc_shr)
 {
 
-    for(int k = 0; k < length; k++){
+    for(size_t k = 0; k < length; k++){
         acc[k] = vlashr16(acc[k], acc_shr);
         const vpu_int16_acc_t tmp = vlmacc16(0, b[k], c[k]);
         acc[k] = vlsub16(acc[k], vlsat16(tmp, bc_shr));
@@ -61,7 +61,7 @@ headroom_t vect_s32_macc(
     const right_shift_t c_shr)
 {
 
-    for(int k = 0; k < length; k++){
+    for(size_t k = 0; k < length; k++){
         acc[k] = vlashr32(acc[k], acc_shr);
         const int32_t B = vlashr32(b[k], b_shr);
         const int32_t C = vlashr32(c[k], c_shr);
@@ -82,7 +82,7 @@ headroom_t vect_s32_nmacc(
     const right_shift_t c_shr)
 {
 
-    for(int k = 0; k < length; k++){
+    for(size_t k = 0; k < length; k++){
         acc[k] = vlashr32(acc[k], acc_shr);
         const int32_t B = vlashr32(b[k], b_shr);
         const int32_t C = vlashr32(c[k], c_shr);
